+++ super_editor/lib/super_editor.dart
@@ -87,7 +87,9 @@ export 'src/infrastructure/touch_controls.dart';
 export 'src/infrastructure/text_input.dart';
 export 'src/infrastructure/viewport_size_reporting.dart';
 export 'src/infrastructure/popovers.dart';
-export 'src/infrastructure/dropdown.dart';
+export 'src/infrastructure/popover_scaffold.dart';
+export 'src/infrastructure/selectable_list.dart';
+export 'src/infrastructure/default_popovers.dart';
 
 // Super Reader
 export 'src/super_reader/read_only_document_android_touch_interactor.dart';
